dockerfiles/web_app_django/src/requirements.txt
django-rest-framework
mysql-connector-python
boto3
